moved ai call to services area
